backend/requirements.txt
fastapi
uvicorn
supabase
openai
google-generativeai
python-dotenv
requests
beautifulsoup4
feedparser
pydantic
